# Optimize pileup calling for high throughput targeted amplicon bed files